pyproject.toml
[project]
name = "telegram-bot-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.26.0",
    "aiohttp-socks>=0.11.0",
    "aiosqlite>=0.21.0",
    "loguru>=0.7.3",
    "pillow>=12.1.1",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "sqlalchemy[asyncio]>=2.0.39",
    "sqlmodel>=0.0.24",
    "uvloop>=0.22.1 ; sys_platform != 'win32'",
    "winloop>=0.3.1 ; sys_platform == 'win32'",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E", # pycodestyle 错误检查
    "W", # pycodestyle 警告检查
    "F", # pyflakes 错误检查
    "I", # isort 导入排序
    "B", # flake8-bugbear 常见错误检查
    "C4", # flake8-comprehensions 列表/字典推导式检查
    "UP", # pyupgrade 自动升级语法
]
ignore = [
    "B008", # 不在参数默认值中执行函数调用
    "C901", # 函数复杂度过高
]

